courseapp/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-ckeditor==6.7.1
django-debug-toolbar==4.3.0
django-js-asset==2.2.0
pillow==10.2.0
PyMySQL==1.1.0
sqlparse==0.4.4
tzdata==2024.1
